+++ internal/infra/push/datadestination_mariadb.go
@@ -81,7 +81,7 @@ func (d MariadbDialect) InsertStatement(tableName string, columns []string, valu
 	return fmt.Sprintf("INSERT IGNORE INTO %s(%s) VALUES(%s)", tableName, strings.Join(protectedColumns, ","), strings.Join(values, ","))
 }
 
-func (d MariadbDialect) UpdateStatement(tableName string, columns []string, uValues []string, primaryKeys []string, pValues []string) (string, *push.Error) {
+func (d MariadbDialect) UpdateStatement(tableName string, columns []string, uValues []string, primaryKeys []string, pValues []string) (string, []string, *push.Error) {
 	sql := &strings.Builder{}
 	sql.Write([]byte("UPDATE "))
 	sql.Write([]byte(tableName))
@@ -97,7 +97,7 @@ func (d MariadbDialect) UpdateStatement(tableName string, columns []string, uVal
 	if len(primaryKeys) > 0 {
 		sql.Write([]byte(" WHERE "))
 	} else {
-		return "", &push.Error{Description: fmt.Sprintf("can't update table [%s] because no primary key is defined", tableName)}
+		return "", []string{}, &push.Error{Description: fmt.Sprintf("can't update table [%s] because no primary key is defined", tableName)}
 	}
 	for index, pk := range primaryKeys {
 		sql.Write([]byte(pk))
@@ -107,7 +107,7 @@ func (d MariadbDialect) UpdateStatement(tableName string, columns []string, uVal
 			sql.Write([]byte(" AND "))
 		}
 	}
-	return sql.String(), nil
+	return sql.String(), append(columns, primaryKeys...), nil
 }
 
 // IsDuplicateError check if error is a duplicate error