+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Types of changes
 - `Added` configuration of exported columns in tables.yaml, see issue #33 for more information.
 - `Added` MariaDB/MySQL support (thanks to @joaking85).
 - `Changed` order of keys in output JSON lines will be alphabetical when pulling (without configuration in tables.yaml).
+- `Fixed` limit keyword on DB2 dialect.
 
 ## [1.8.0]